Push New Method

+++ src/main/java/com/example/gestionbibliotheque/bean/Reception.java
@@ -14,13 +14,13 @@ public class Reception {
     private long id ;
     private LocalDateTime dateReception;
     private String reference;
-    private double quantitereceptione;
+    private int quantitereceptione;
 
-    public double getQuantitereceptione() {
+    public int getQuantitereceptione() {
         return quantitereceptione;
     }
 
-    public void setQuantitereceptione(double quantitereceptione) {
+    public void setQuantitereceptione(int quantitereceptione) {
         this.quantitereceptione = quantitereceptione;
     }
